move header files to libshell folder
makes them easier to disambiguate when including from an external project
