+++ include/libshell/BilayerStVKMaterial.h
@@ -53,6 +53,6 @@ namespace LibShell {
 
 
     };
-};
+}
 
 #endif
