+++ include/libshell/MaterialModel.h
@@ -29,6 +29,6 @@ namespace LibShell {
             Eigen::Matrix<double, 1, 18 + 3 * SFF::numExtraDOFs>* derivative, // F(face, i), then the three vertices opposite F(face,i), then the extra DOFs on oppositeEdge(face,i)
             Eigen::Matrix<double, 18 + 3 * SFF::numExtraDOFs, 18 + 3 * SFF::numExtraDOFs>* hessian) const = 0;
     };
-};
+}
 
 #endif
