+++ include/libshell/MidedgeAngleSinFormulation.h
@@ -1,5 +1,5 @@
-#ifndef MIDEDGEANGLETANFORMULATION_H
-#define MIDEDGEANGLETANFORMULATION_H
+#ifndef MIDEDGEANGLESINFORMULATION_H
+#define MIDEDGEANGLESINFORMULATION_H
 
 #include <Eigen/Core>
 #include <vector>
@@ -9,10 +9,11 @@ namespace LibShell {
     class MeshConnectivity;
 
     /*
-    * Second fundamental form based on rotating the averaged face normals across edges by an addition per-edge angle.
-    * Uses the tan discretization of curvature, which diverges for collapsed hinges
-    */
-    class MidedgeAngleTanFormulation
+     * Second fundamental form based on rotating the averaged face normals across edges by an addition per-edge angle.
+     * Uses the sin discretization of curvature.
+     */
+
+    class MidedgeAngleSinFormulation
     {
     public:
         constexpr static int numExtraDOFs = 1;
@@ -25,8 +26,8 @@ namespace LibShell {
             const Eigen::VectorXd& extraDOFs,
             int face,
             Eigen::Matrix<double, 4, 18 + 3 * numExtraDOFs>* derivative, // F(face, i), then the three vertices opposite F(face,i), then the thetas on oppositeEdge(face,i)
-            std::vector<Eigen::Matrix<double, 18 + 3 * numExtraDOFs, 18 + 3 * numExtraDOFs> >* hessian);
+            std::vector < Eigen::Matrix<double, 18 + 3 * numExtraDOFs, 18 + 3 * numExtraDOFs> >* hessian);
     };
-};
+}
 
 #endif
