+++ include/libshell/TensionFieldStVKMaterial.h
@@ -1,25 +1,27 @@
-#ifndef STVKMATERIAL_H
-#define STVKMATERIAL_H
+#ifndef TENSIONFIELDSTVKMATERIAL_H
+#define TENSIONFIELDSTVKMATERIAL_H
 
 #include "MaterialModel.h"
 
 namespace LibShell {
 
     /*
-     * St. Venant-Kirchhoff linear material model, with energy density
+     * Tension field theory implementation of a St. Venant-Kirchhoff material with
+     * energy density (in the pure-tension case)
      * W = alpha/2.0 tr(S)^2 + beta tr(S^2),
      * for strain tensor S = gbar^{-1}(g-gbar), where g and gbar are the current
      * and rest metrics of the shell volume (which vary in the thickness direction
      * as defined by the surface fundamental forms).
      *
-     * Needs a MonolayerRestState.
+     * This material has no bending energy.
+     * Takes a MonolayerRestState.  
      */
 
     template <class SFF>
-    class StVKMaterial : public MaterialModel<SFF>
+    class TensionFieldStVKMaterial : public MaterialModel<SFF>
     {
     public:
-        StVKMaterial(double lameAlpha, double lameBeta) : lameAlpha_(lameAlpha), lameBeta_(lameBeta) {}
+        TensionFieldStVKMaterial(double lameAlpha, double lameBeta) : lameAlpha_(lameAlpha), lameBeta_(lameBeta) {}
 
         /*
          * Lame parameters of the material (as in the energy density written above)
@@ -42,9 +44,7 @@ namespace LibShell {
             int face,
             Eigen::Matrix<double, 1, 18 + 3 * SFF::numExtraDOFs>* derivative, // F(face, i), then the three vertices opposite F(face,i), then the extra DOFs on oppositeEdge(face,i)
             Eigen::Matrix<double, 18 + 3 * SFF::numExtraDOFs, 18 + 3 * SFF::numExtraDOFs>* hessian) const;
-
-
     };
-};
+}
 
 #endif
