+++ src/GeometryDerivatives.cpp
@@ -1,8 +1,7 @@
 #include "GeometryDerivatives.h"
-#include "../include/MeshConnectivity.h"
+#include "libshell/MeshConnectivity.h"
 #include <iostream>
 #include <random>
-#include <Eigen/Geometry>
 
 namespace LibShell {
 
@@ -230,4 +229,4 @@ namespace LibShell {
         return result;
     }
 
-};
+}
