+++ src/MaterialModel/TensionFieldStVKMaterial.cpp
@@ -1,12 +1,12 @@
-#include "../../include/TensionFieldStVKMaterial.h"
-#include "../../include/MeshConnectivity.h"
-#include <vector>
+#include "libshell/TensionFieldStVKMaterial.h"
+#include "libshell/MeshConnectivity.h"
+#include "libshell/MidedgeAngleSinFormulation.h"
+#include "libshell/MidedgeAngleTanFormulation.h"
+#include "libshell/MidedgeAverageFormulation.h"
+#include "libshell/RestState.h"
 #include "../GeometryDerivatives.h"
 #include <Eigen/Dense>
-#include "../../include/MidedgeAngleSinFormulation.h"
-#include "../../include/MidedgeAngleTanFormulation.h"
-#include "../../include/MidedgeAverageFormulation.h"
-#include "../../include/RestState.h"
+#include <vector>
 
 namespace LibShell {
 
@@ -197,4 +197,4 @@ namespace LibShell {
     template class TensionFieldStVKMaterial<MidedgeAngleTanFormulation>;
     template class TensionFieldStVKMaterial<MidedgeAverageFormulation>;
 
-};
+}
