+++ tests/main.cpp
@@ -1,18 +1,18 @@
+#include "findiff.h"
+#include "libshell/BilayerStVKMaterial.h"
+#include "libshell/ElasticShell.h"
+#include "libshell/MeshConnectivity.h"
+#include "libshell/MidedgeAngleTanFormulation.h"
+#include "libshell/MidedgeAngleSinFormulation.h"
+#include "libshell/MidedgeAverageFormulation.h"
+#include "libshell/NeoHookeanMaterial.h"
+#include "libshell/RestState.h"
+#include "libshell/StVKMaterial.h"
+#include "libshell/TensionFieldStVKMaterial.h"
+#include <cmath>
 #include <Eigen/Core>
 #include <iostream>
 #include <map>
-#include <cmath>
-#include "../include/MeshConnectivity.h"
-#include "../include/ElasticShell.h"
-#include "../include/MidedgeAngleTanFormulation.h"
-#include "../include/MidedgeAngleSinFormulation.h"
-#include "../include/MidedgeAverageFormulation.h"
-#include "../include/StVKMaterial.h"
-#include "../include/BilayerStVKMaterial.h"
-#include "../include/TensionFieldStVKMaterial.h"
-#include "../include/NeoHookeanMaterial.h"
-#include "../include/RestState.h"
-#include "findiff.h"
 #include <random>
 
 std::default_random_engine rng;
